Fix Clippy warnings

+++ crates/rmd-cli/src/format_helpers.rs
@@ -90,7 +90,7 @@ pub fn format_ls_time(rfc3339: &str) -> String {
 fn days_from_civil(y: i64, m: u32, d: u32) -> i64 {
     let y = if m <= 2 { y - 1 } else { y };
     let era = y.div_euclid(400);
-    let yoe = (y - era * 400) as i64;
+    let yoe = y - era * 400;
     let m = m as i64;
     let d = d as i64;
     let doy = (153 * (if m > 2 { m - 3 } else { m + 9 }) + 2) / 5 + d - 1;
@@ -106,7 +106,7 @@ fn civil_from_epoch(epoch: i64) -> (i64, u32, u32, u32, u32) {
 
     let z = days + 719_468;
     let era = z.div_euclid(146_097);
-    let doe = (z - era * 146_097) as i64;
+    let doe = z - era * 146_097;
     let yoe = (doe - doe / 1_460 + doe / 36_524 - doe / 146_096) / 365;
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
